flask-app/requirements.txt
backpack==0.1
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
cleo==0.6.8
Click==7.0
dominate==2.3.4
Faker==0.8.18
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
orator==0.9.8
pastel==0.1.0
pendulum==1.5.1
psycopg2==2.7.5
pyaml==16.12.2
Pygments==2.2.0
pylev==1.3.0
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
six==1.11.0
text-unidecode==1.2
tzlocal==1.5.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
